python_ai/requirements.txt

Flask==2.3.3
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
SpeechRecognition==3.10.0
torch==2.0.1
torchvision==0.15.2
transformers==4.34.0
